// FRONT/src/app/interfaces/origination/pre-pin/pre-pin.interface.ts
export interface PrePin {
    Id_pre_pin_assumptions:                   number;
    project_id:                               number;
    descriptions:                             string;
    id_certification_program:                 number;
    id_registration_route:                    number;
    id_methodology:                           number;
    project_area:                             string;
    id_project_condition:                     number;
    id_licences_permits:                      number;
    id_confidence_of_crediting_activity_area: number;
    id_umafor:                                number;
    authorized_harvesting_volume:             number;
    underlying_activities:                    string;
    id_estimate_permanence:                   number;
    estimate_sample_size:                     string;
    id_estimate_mrv:                          number;
    id_estimate_leakage:                      number;
    id_estimate_reversal_risk:                number;
    percentage_mki_price:                     number;
    conficence_upfront_costs:                 number;
    cba:                                      number;
    irr:                                      number;
    credit_type:                              string;
    id_safeguards:                            number;
    id_social_community:                      number;
    id_shareholders:                          number;
    id_press_negative:                        number;
    id_biodiversity:                          number;
    id_human_rights:                          number;
    id_indigenous_people:                     number;
    id_h_s:                                   number;
    id_negative_ehs:                          number;
}