// FRONT/src/app/interfaces/origination/pre-pin/pre-pin-catalogs.interface.ts
export interface Programme {
  Id_programme: number;
  programme:    string;
}

export interface RegistrationRoute {
  Id_registration_route: number;
  registration_route:    string;
}

export interface Methodology {
  Id_methodology: number;
  methodology:    string;
}

export interface ProjectCondition {
  Id_project_condition: number;
  project_condition:    string;
}

export interface LicensesPermits {
  Id_licenses_permits: number;
  licenses_permits:    string;
}

export interface ConfidenceCreditingActivityArea {
  Id_confidence_crediting_activity_area: number;
  confidence_crediting_activity_area:    string;
}

export interface EstimatePermanence {
  Id_estimate_permanence: number;
  estimate_permanence:    string;
}

export interface EstimatedMrvRequirements {
  Id_estimated_mrv_requirements: number;
  estimated_mrv_requirements:    string;
}

export interface CbaCalculatorVersion {
  Id_cba_calculator_version: number;
  cba_calculator_version:    string;
}

export interface EstimateLeakage {
  Id_estimate_leakage: number;
  estimate_leakage:    string;
}

export interface SafeguardsNoHarmApproach {
  Id_safeguards_no_harm_approach: number;
  safeguards_no_harm_approach:    string;
}

export interface SocialCommunityNoHarm {
  Id_social_community_no_harm: number;
  social_community_no_harm:    string;
}

export interface ShareholdersEngagement {
  Id_shareholders_engagement: number;
  shareholders_engagement:    string;
}

export interface Biodiversity {
  Id_biodiversity: number;
  biodiversity:    string;
}

export interface Hs {
  Id_hs: number;
  hs:    string;
}

export interface NegativeEhs {
  Id_negative_ehs: number;
  negative_ehs:    string;
}

export interface IndigenousPeople {
  Id_indigenouspeople: number;
  indigenous_people:   string;
}

export interface HumanRights {
  Id_human_rights: number;
  human_rights:    string;
}

export interface BufferPool {
  Id_buffer_pool:   number;
  buffer:           string;
}

export interface InventoryStratification {
  Id_inventory_stratification: number;
  inventory_stratification:    string;
}

export interface PressNegative {
  Id_press_negative: number;
  press_negative:    string;
}

export interface GrievanceMechanism {
  Id_grievance_mechanism: number;
  grievance_mechanism:    string;
}